Move some stuff to new gui utils module
